% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_default}
\alias{apply_default}
\title{Apply default values}
\usage{
apply_default(x, default)
}
\arguments{
\item{x}{Vector.}

\item{default}{Vector of default values, matched by position.}
}
\description{
Apply default values
}
\examples{
default <- c(1, 2, 3, 4)
x <- c(10)
apply_default(x, default)
x <- c(10, NA, 30)
apply_default(x, default)
}
