% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_table}
\alias{as_table}
\title{Coerce to table}
\usage{
as_table(x, type = data.frame(), ...)
}
\arguments{
\item{x}{Object passed to the corresponding table coercion function.}

\item{type}{Table type (passed to \code{\link{parse_table_type}}).}

\item{...}{Arguments passed to the corresponding table coercion function.}
}
\description{
Coerces an object to a table of the desired type.
}
\examples{
x <- list(a = 1, b = 1)
as_table(x)
\dontrun{
as_table(x, "data.table")
as_table(x, "tibble")
as_table(x, "tbl_dt")
}
}
\seealso{
Other Table functions: \code{\link{create_table}},
  \code{\link{is_table}}, \code{\link{parse_table_type}},
  \code{\link{parse_table}},
  \code{\link{remove_empty_dimensions}},
  \code{\link{write_table}}
}
