% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.R
\name{as_time}
\alias{as_time}
\title{Convert to datetime}
\usage{
as_time(x, origin = as.POSIXct("1970-01-01", tz = "UTC"), tz = "UTC",
  format = NULL)
}
\arguments{
\item{x}{Object to be converted.}

\item{origin}{Date-time object, or something which can be coerced by as.POSIXct(tz = "GMT") to such an object.}

\item{tz}{A time zone specification to be used for the conversion.}

\item{format}{Character string giving a date-time format as used by \code{\link[base]{strptime}}.}
}
\description{
Convert to datetime
}
\examples{
x <- Sys.time()
x == as_time(as.numeric(x))
x <- 1:10
as_time(x)
x <- data.frame(t1 = 1:10, t2 = 1:10)
as_time(x)
}
\seealso{
Other temporal functions: \code{\link{as_time_doy}}
}
