% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.R
\name{as_time_doy}
\alias{as_time_doy}
\title{Convert day-of-year to datetime}
\usage{
as_time_doy(year, day, utc_offset = 0)
}
\arguments{
\item{year}{Year.}

\item{day}{Day of year.}

\item{utc_offset}{Offset (in hours) of local time relative to UTC.}
}
\description{
Convert day-of-year to datetime
}
\examples{
as_time_doy(1984, 214, -8) == as_time("1984-08-01 08:00:00", format = "\%F \%T")
}
\seealso{
Other temporal functions: \code{\link{as_time}}
}
