% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{as_units}
\alias{as_units}
\title{Coerce to units}
\usage{
as_units(x, unit = units::unitless)
}
\arguments{
\item{x}{Object.}

\item{unit}{Units or character string (see \code{\link{parse_unit}}).}
}
\description{
Converts an object to \code{units} using \code{\link[units]{as.units}}.
}
\examples{
as_units(1:5, "m/s")
as_units(1:5, parse_unit("m/s"))
}
\seealso{
Other Unit functions: \code{\link{convert_units}},
  \code{\link{is_units}}, \code{\link{parse_unit}}
}
