% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{bisq}
\alias{bisq}
\title{Bisquare kernel}
\usage{
bisq(dx, width)
}
\arguments{
\item{dx}{Distances from center.}

\item{width}{Width of kernel.}
}
\description{
Bisquare kernel
}
\examples{
bisq(-6:6, 10)
}
