% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{colranges}
\alias{colranges}
\title{Generate ranges from a column (vector)}
\usage{
colranges(t, maxdt = Inf, x = NULL)
}
\arguments{
\item{t}{(atomic) Measures along a single dimension (temporal or otherwise) coercible to numeric.}

\item{maxdt}{(numeric) Maximum difference in \code{t} to consider as continuous.}

\item{x}{(atomic) Values of the same length as \code{t} used to ignore missing values (\code{NA}).}
}
\description{
Constructs ranges that fit the coverage of point measures.
}
