% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_length.R
\name{count_true}
\alias{count_true}
\title{Cumulative count of consecutive TRUE}
\usage{
count_true(x)
}
\arguments{
\item{x}{Logical vector.}
}
\description{
Cumulative count of consecutive TRUE
}
\examples{
x <- c(TRUE, TRUE, FALSE, TRUE)
count_true(x)
}
