% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_table}
\alias{create_table}
\title{Create a table}
\usage{
create_table(..., type = data.frame(), stringsAsFactors = FALSE)
}
\arguments{
\item{...}{Arguments passed to the corresponding table creation function.}

\item{type}{Table type (passed to \code{\link{parse_table_type}}).}

\item{stringsAsFactors}{Whether to convert all character columns to factors.}
}
\description{
Creates a table of the desired type.
}
\examples{
create_table(a = 1, b = 2)
\dontrun{
create_table(a = 1, b = 2, type = "data.table")
create_table(a = 1, b = 2, type = "tibble")
create_table(a = 1, b = 2, type = "tbl_dt")
}
}
\seealso{
Other Table functions: \code{\link{as_table}},
  \code{\link{is_table}}, \code{\link{parse_table_type}},
  \code{\link{parse_table}}, \code{\link{rbind_tables}},
  \code{\link{remove_empty_dimensions}},
  \code{\link{write_table}}
}
