% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{datapackage_col_type}
\alias{datapackage_col_type}
\title{Create column specification}
\usage{
datapackage_col_type(type, format = "")
}
\arguments{
\item{type}{Column type.}

\item{format}{Format specification (e.g. see \code{\link[readr]{parse_datetime}}). If set to "", date times are parsed as ISO8601, and dates and times are parsed using the date and time formats specified in the \code{\link[readr]{locale}}.}
}
\description{
Return the \code{readr::\link[readr]{cols}} column specification for the Data Package Table Schema column type.
}
\examples{
datapackage_col_type("string")
datapackage_col_type("number")
datapackage_col_type("datetime")
datapackage_col_type("boolean")
}
\seealso{
Other Data package functions: \code{\link{datapackage_data}},
  \code{\link{datapackage_info}}
}
