% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{datapackage_data}
\alias{datapackage_data}
\alias{datapackage_data_github}
\title{Read Data Package data}
\usage{
datapackage_data(path = ".", resources = NULL, simplify = TRUE)

datapackage_data_github(repo, resources = NULL, simplify = TRUE)
}
\arguments{
\item{path}{Character string.}

\item{resources}{Resources to read, by name or relative path.}

\item{simplify}{Whether to return single results unlisted.}

\item{repo}{GitHub repository address (see \code{\link{github_raw_url}}).}
}
\description{
NOTE: Only Table Schema and CSV files are supported.
}
\examples{
datapackage_data_github("columbia-glacier/optical-surveys-1985")
datapackage_data_github("columbia-glacier/noaa-coops", resources = "columbia-glacier")
}
\seealso{
Other Data package functions: \code{\link{datapackage_col_type}},
  \code{\link{datapackage_info}}
}
