% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packages.R
\name{datapackage_info}
\alias{datapackage_info}
\alias{datapackage_info_github}
\title{Read Data Package metadata}
\usage{
datapackage_info(path = ".", simplifyVector = TRUE, ...)

datapackage_info_github(repo, simplifyVector = TRUE, ...)
}
\arguments{
\item{path}{Character string.}

\item{simplifyVector}{Whether to simplify nested lists into vectors and data frames.}

\item{...}{Arguments passed to \code{\link[jsonlite]{fromJSON}}.}

\item{repo}{GitHub repository address (see \code{\link{github_raw_url}}).}
}
\description{
Read Data Package metadata
}
\examples{
datapackage_info_github("columbia-glacier/noaa-coops")
}
\seealso{
Other Data package functions: \code{\link{datapackage_col_type}},
  \code{\link{datapackage_data}}
}
