% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{res}
\alias{res}
\alias{ssq}
\alias{mse}
\alias{rmse}
\alias{mae}
\title{Error functions}
\usage{
res(sim, obs)

ssq(sim, obs)

mse(sim, obs)

rmse(sim, obs)

mae(sim, obs)
}
\arguments{
\item{sim}{Simulated values.}

\item{obs}{Observed values.}
}
\description{
\describe{
  \item{res}{Residual}
  \item{ssq}{Sum of squares}
  \item{mse}{Mean squared error}
  \item{rmse}{Root-mean-square error}
  \item{mae}{Mean absolute error}
}
}
\examples{
sim <- 1:2
obs <- numeric(2)
res(sim, obs)
ssq(sim, obs)
mse(sim, obs)
rmse(sim, obs)
mae(sim, obs)
}
