% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{eval_parser}
\alias{eval_parser}
\title{Evaluate a parser}
\usage{
eval_parser(parser, envir = parent.frame(), enclos = globalenv(),
  default = NULL)
}
\arguments{
\item{parser}{Object to be evaluated.}

\item{envir}{Environment in which \code{parser} is to be evaluated. Must contain objects whose names are in \code{(attr(parser, "required")}.}

\item{enclos}{Environment in which to look for objects not found in \code{envir}.}

\item{default}{Default value (or function) to return (or evaluate) if requirements are not met.}
}
\description{
Evaluates an unevaluated expression after first checking that the required object names are present in the specified environment.
}
\examples{
p <- parser(x + 10, required = "x")
eval_parser(p, envir = list(x = 1))
eval_parser(p, envir = list(y = 1))
\dontrun{
eval_parser(p, envir = list(y = 1), default = function() stop("Requirements not met"))
}
}
\seealso{
Other Parser functions: \code{\link{parser}}
}
