% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{feval}
\alias{feval}
\title{Evaluate a function}
\usage{
feval(fun, envir = parent.frame(), enclos = globalenv(), default = NULL)
}
\arguments{
\item{fun}{Fuction to evaluate.}

\item{envir}{Environment in which \code{fun} is to be evaluated. Must contain objects of the same name as the arguments in \code{f}.}

\item{enclos}{Environment in which to look for objects not found in \code{envir}. Relevant only when \code{envir} is a (pair)list or data frame.}

\item{default}{Default value (or function) to return (or evaluate) if requirements are not met.}
}
\description{
Evaluates a function by taking its argument values by name from the specified environment.
}
\examples{
f <- function(x) {x + 1}
feval(f, envir = list(x = 1))
feval(f, envir = list(y = 1))
\dontrun{
feval(f, envir = list(y = 1), default = function() stop("Arguments not found"))
}
}
