% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{find_interval}
\alias{find_interval}
\title{Find interval indices}
\usage{
find_interval(pos, breaks)
}
\arguments{
\item{pos}{(numeric)}

\item{breaks}{(numeric) Sorted (weakly) increasingly.}
}
\description{
Find interval indices
}
