% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{github_raw_url}
\alias{github_raw_url}
\title{GitHub Raw URL from Repository Address}
\usage{
github_raw_url(repo)
}
\arguments{
\item{repo}{Repository address in the format \code{username/repo[/subdir][@ref]}, where \code{ref} can be a commit, tag, or branch name. Defaults to "master".}
}
\description{
Get the URL to the raw GitHub user content corresponding to a GitHub repository address.
}
\examples{
github_raw_url("username/repo")
github_raw_url("username/repo/subdir")
github_raw_url("username/repo/subdir@ref")
}
