% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_length.R
\name{group_true}
\alias{group_true}
\title{Group number of consecutive TRUE}
\usage{
group_true(x)
}
\arguments{
\item{x}{Logical vector.}
}
\description{
Group number of consecutive TRUE
}
\examples{
x <- c(TRUE, TRUE, FALSE, TRUE)
group_true(x)
}
