% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_length.R
\name{length_true}
\alias{length_true}
\title{Length of consecutive TRUE}
\usage{
length_true(x)
}
\arguments{
\item{x}{Logical vector.}
}
\description{
Length of consecutive TRUE
}
\examples{
x <- c(TRUE, TRUE, FALSE, TRUE)
length_true(x)
}
