% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{nearest_distance}
\alias{nearest_distance}
\title{Distance to nearest element}
\usage{
nearest_distance(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Distance to nearest element
}
\examples{
x <- c(0, 1, 5, 9)
nearest_distance(x)
}
