% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{nearest_width}
\alias{nearest_width}
\title{Width of bounding midpoints}
\usage{
nearest_width(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Width of bounding midpoints
}
\examples{
x <- c(0, 1, 5, 9)
nearest_width(x)
}
