% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{parse_table}
\alias{parse_table}
\title{Parse a table}
\usage{
parse_table(x, parsers, as = c("table", "list"), ...)
}
\arguments{
\item{x}{Table.}

\item{parsers}{Parser or list thereof (see \code{\link{parser}}).}

\item{as}{Type of object to return. If "table", a table of the same type as \code{x}, otherwise a list.}

\item{...}{Arguments passed to \code{\link{eval_parser}}.}
}
\description{
Applies parsers to a table and coerces the result to a table of the same type.
}
\examples{
x <- data.frame(x = 1:5, y = 6:10)
parsers <- list(
  xx = parser(x + 10, "x"),
  xy = parser(x + y, c("x", "y"))
)
parse_table(x, parsers)
parse_table(x, parsers, as = "table")
parsers$z = parser(z * 10, "z")
parse_table(x, parsers)
parse_table(list(x = 1:5), parsers)
}
\seealso{
Other Table functions: \code{\link{as_table}},
  \code{\link{create_table}}, \code{\link{is_table}},
  \code{\link{parse_table_type}},
  \code{\link{remove_empty_dimensions}}
}
