% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{parse_table_type}
\alias{parse_table_type}
\title{Parse table type}
\usage{
parse_table_type(x)
}
\arguments{
\item{x}{Table or character string.}
}
\description{
Parses a table or character string into the corresponding table type.
}
\examples{
parse_table_type(data.frame())
parse_table_type("data.frame")
parse_table_type("data.table")
parse_table_type("tibble")
parse_table_type("tbl_df")
parse_table_type("tbl_dt")
\dontrun{
parse_table_type("fake.table")
}
\dontrun{
parse_table_type(data.table::data.table())
parse_table_type(tibble::tibble())
parse_table_type(dtplyr::tbl_dt(data.table::data.table()))
}
}
\seealso{
Other Table functions: \code{\link{as_table}},
  \code{\link{create_table}}, \code{\link{is_table}},
  \code{\link{parse_table}},
  \code{\link{remove_empty_dimensions}}
}
