% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{parse_unit}
\alias{parse_unit}
\title{Parse unit}
\usage{
parse_unit(x)
}
\arguments{
\item{x}{Character string.}
}
\description{
Parses a character string as a unit of measure according to the following notations:
\enumerate{
  \item Symbolic (e.g. "m/s") via \code{\link[units]{ud_units}}
  \item Product power (e.g. "m s-1") via \code{\link[units]{parse_unit}}
}
}
\details{
For a list of supported units and their notation, see the UDUNITS-2 \href{https://www.unidata.ucar.edu/software/udunits/udunits-2.2.25/doc/udunits/udunits2.html#Database}{Units Database}.
}
\examples{
parse_unit("m/s")
parse_unit("m s-1")
parse_unit("kg * (m^2 / s^3)")
parse_unit("kg m2 s-3")
}
\seealso{
Other Unit functions: \code{\link{as_units}},
  \code{\link{convert_units}}, \code{\link{is_units}}
}
