% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser}
\alias{parser}
\title{Create a parser}
\usage{
parser(expr, required = NULL)
}
\arguments{
\item{expr}{Any syntactically valid \code{R} expression.}

\item{required}{Character vector of required object names.}
}
\description{
Creates an unevaluated expression with an attribute for required object names.
}
\examples{
parser(x + 10, required = "x")
parser(x + y, required = c("x", "y"))
}
\seealso{
Other Parser functions: \code{\link{eval_parser}}
}
