% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{paste2}
\alias{paste2}
\title{Concatenate Strings}
\source{
\url{http://stackoverflow.com/a/31508774}
}
\usage{
paste2(..., sep = " ", collapse = NULL, na.rm = FALSE)
}
\arguments{
\item{...}{Objects converted to character vectors.}

\item{sep}{Character string to seperate the terms.}

\item{collapse}{Character string to separate the results.}

\item{na.rm}{Whether \code{NA} values should be skipped.}
}
\description{
Extends \code{\link[base]{paste}} with the ability to remove missing values (\code{NA}).
}
\examples{
paste2(c("hello", "goodbye"), NA, NA, "world", NA)
paste2(c("hello", "goodbye"), NA, NA, "world", NA, na.rm = TRUE)
}
