% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{range_mean}
\alias{range_mean}
\title{Mean of range endpoints}
\usage{
range_mean(ranges)
}
\arguments{
\item{ranges}{Ranges.}
}
\description{
Mean of range endpoints
}
\examples{
ranges <- as_time(cbind(0, 10))
range_mean(ranges)
}
