% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{range_select}
\alias{range_select}
\title{Select elements within ranges}
\usage{
range_select(x, ranges)
}
\arguments{
\item{x}{Position vector to evaluate against ranges.}

\item{ranges}{Ranges.}
}
\description{
Select elements within ranges
}
\examples{
x <- as_time(1:10)
ranges <- x[1:3]
range_select(x, ranges)
ranges <- as_time(data.frame(1, 2:4))
range_select(x, ranges)
}
