% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{range_split}
\alias{range_split}
\title{Split object by ranges}
\usage{
range_split(x, t, ranges)
}
\arguments{
\item{x}{Vector or data frame to be split into groups.}

\item{t}{Position vector to evaluate against ranges.}

\item{ranges}{Ranges.}
}
\description{
Split object by ranges
}
\examples{
x <- 1:10
t <- as_time(1:10)
ranges <- t[1:3]
range_split(x, t, ranges)
ranges <- as_time(data.frame(1, 2:4))
range_split(x, t, ranges)
x <- data.frame(V1 = 1:10, V2 = 1:10)
range_split(x, t, ranges)
}
