% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_to_datetime}
\alias{ranges_to_datetime}
\title{Coerce range endpoints to date-time}
\usage{
ranges_to_datetime(ranges, ...)
}
\arguments{
\item{ranges}{Ranges.}

\item{...}{Arguments passed to \code{\link{as_time}}.}
}
\description{
Coerce range endpoints to date-time
}
