% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_to_numeric}
\alias{ranges_to_numeric}
\title{Coerce range endpoints to numeric}
\usage{
ranges_to_numeric(ranges)
}
\arguments{
\item{ranges}{Ranges.}
}
\description{
Coerce range endpoints to numeric
}
