% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rbind_tables}
\alias{rbind_tables}
\title{Bind tables by row}
\usage{
rbind_tables(l, use.names = fill, fill = TRUE, ...)
}
\arguments{
\item{l}{List containing \code{data.table}, \code{data.frame} or \code{list} objects.}

\item{use.names}{Whether to bind tables by matching column names.}

\item{fill}{Whether to fill missing columns with \code{NA}.}

\item{...}{Arguments passed to \code{\link[data.table]{rbindlist}}.}
}
\description{
Same as \code{\link[data.table]{rbindlist}}, but returns a table of the same type as the inputs.
}
\examples{
ldf <- list(
  data.frame(x = 1),
  data.frame(x = 2)
)
str(rbind_tables(ldf))
\dontrun{
ldt <- lapply(ldf, as_table, "data.table")
str(rbind_tables(ldt))
ltb <- lapply(ldf, as_table, "tibble")
str(rbind_tables(ltb))
}
}
\seealso{
Other Table functions: \code{\link{as_table}},
  \code{\link{create_table}}, \code{\link{is_table}},
  \code{\link{parse_table_type}},
  \code{\link{parse_table}},
  \code{\link{remove_empty_dimensions}},
  \code{\link{write_table}}
}
