% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{remove_empty_dimensions}
\alias{remove_empty_dimensions}
\title{Remove empty table dimensions}
\usage{
remove_empty_dimensions(x, ignore = NULL)
}
\arguments{
\item{x}{Table.}

\item{ignore}{Vector of column names to ignore when checking for empty rows.}
}
\description{
Removes empty rows and columns from a table.
}
\examples{
df <- data.frame(x = 1, y = NA)
remove_empty_dimensions(df)
remove_empty_dimensions(df, ignore = "x")
\dontrun{
dt <- data.table::data.table(x = 1, y = NA)
remove_empty_dimensions(dt)
remove_empty_dimensions(dt, ignore = "x")
tb <- tibble::tibble(x = 1, y = NA)
remove_empty_dimensions(tb)
remove_empty_dimensions(tb, ignore = "x")
}
}
\seealso{
Other Table functions: \code{\link{as_table}},
  \code{\link{create_table}}, \code{\link{is_table}},
  \code{\link{parse_table_type}},
  \code{\link{parse_table}}, \code{\link{rbind_tables}},
  \code{\link{write_table}}
}
