% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{roll_ranges}
\alias{roll_ranges}
\title{Generate rolling ranges}
\usage{
roll_ranges(x, dx = NULL, dxi = 1, align = c("left", "center", "right"),
  trim = c("none", "snap", "cut", "drop"))
}
\arguments{
\item{x}{Numeric vector.}

\item{dx}{Width of ranges (by values of \code{x}).}

\item{dxi}{Width of ranges (by number of elements of \code{x}).}

\item{align}{Alignment of ranges with respect to values in \code{x}.}

\item{trim}{Method for trimming ranges extending beyond range of \code{x}.}
}
\description{
Generate rolling ranges
}
\examples{
x <- as_time(seq(0, 60 * 60 * 5, by = 60 * 60))
dx <- 60 * 60 * 2.5
roll_ranges(x, dx, align = "left")
roll_ranges(x, dx, align = "center")
roll_ranges(x, dx, align = "right")
roll_ranges(x, dx, align = "left", trim = "snap")
roll_ranges(x, dx, align = "left", trim = "cut")
roll_ranges(x, dx, align = "left", trim = "drop")
dxi <- 3
roll_ranges(x, dxi = dxi, align = "left")
roll_ranges(x, dxi = dxi, align = "right")
roll_ranges(x, dxi = dxi, align = "center")
}
