% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{sample_interval}
\alias{sample_interval}
\title{Sample variables by interval}
\usage{
sample_interval(.data, pos, breaks, fun = mean, ...)
}
\arguments{
\item{.data}{(data.frame) Variables to sample.}

\item{pos}{(numeric)}

\item{breaks}{(numeric) Sorted (weakly) increasingly.}

\item{fun}{(function|character) Sampling function}

\item{...}{Additional arguments passed to \code{fun}}
}
\description{
Sample variables by interval
}
\examples{
df <- data.frame(
  t = as_time(0:9),
  x = 0:9,
  y = 10:19
)
breaks <- as_time(c(0,5,10))
sample_interval(df[, c("x", "y")], df$t, breaks, fun = mean, na.rm = TRUE)
sample_interval(df[, c("x", "y")], df$t, breaks, fun = max, na.rm = TRUE)
}
