% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{seq_ranges}
\alias{seq_ranges}
\title{Generate sequential ranges}
\usage{
seq_ranges(from = 1, to = from[length(from)], by = (to -
  from[1])/length.out, length.out = as.numeric(to - from[1]), step = by)
}
\arguments{
\item{from}{Starting value of the sequence. If \code{length() > 1}, \code{from = min()} and \code{to = max()}.}

\item{to}{If \code{length(from) = 1}, maximal end value of the sequence.}

\item{by}{Range width.}

\item{length.out}{Desired length of the sequence.}

\item{step}{Increment of the range starting value.}
}
\description{
Generate sequential ranges
}
\examples{
x <- as_time(1:10)
seq_ranges(x, by = 3)
seq_ranges(x, length.out = 2)
seq_ranges(x, by = 3, step = 1)
seq_ranges(x[1], x[3])
}
