% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sim_error}
\alias{sim_error}
\title{Compute errors of simulation}
\usage{
sim_error(par, sim_fun, ..., obs, err_fun = rmse)
}
\arguments{
\item{par}{Parameters of \code{sim_fun}.}

\item{sim_fun}{Simulation function.}

\item{...}{Additional arguments passed to \code{sim_fun}.}

\item{obs}{Observed values.}

\item{err_fun}{Error function.}
}
\description{
Compute errors of simulation
}
\examples{
sim_fun <- function(par) {

}
}
