% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{sp_transform}
\alias{sp_transform}
\title{Transform coordinates}
\usage{
sp_transform(xy, from, to, cols = 1:2)
}
\arguments{
\item{xy}{Data.frame or numeric vector (coerced to single-row data.frame).}

\item{from}{Current proj.4 string.}

\item{to}{Target proj.4 string.}

\item{cols}{Column numbers or names specifying which \code{xy} columns are x and y coordinates.}
}
\description{
For transformations involving NAD27 coordinates ("+datum=NAD27"), custom parameters for Alaska (excluding Aleutian islands) are automatically applied ("+towgs84=-5,135,172"). See \url{http://web.archive.org/web/20130905025856/http://surveying.wb.psu.edu/sur351/DatumTrans/datum_transformations.htm}.
}
\examples{
xy <- c(x = -147.25079, y = 61.18586, z = 505)
sp_transform(xy, "+proj=longlat +datum=WGS84", "+proj=utm +zone=6")
xy <- data.frame(z = 505, x = -147.25079, y = 61.18586)
sp_transform(xy, "+proj=longlat +datum=WGS84", "+proj=utm +zone=6", cols = c("x", "y"))
}
