% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{tblranges}
\alias{tblranges}
\title{Generate ranges from a table}
\usage{
tblranges(.data, time_col, group_col = NULL, maxdt = Inf)
}
\arguments{
\item{.data}{(data.frame)}

\item{time_col}{(character) Name of temporal column.}

\item{group_col}{(character) Name of grouping column.}

\item{maxdt}{(numeric) Maximum difference in \code{time_col} to consider as continuous.}
}
\description{
Like \code{\link{colranges}}, but vectorized over a table.
}
