% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim_list}
\alias{trim_list}
\title{Trim list}
\usage{
trim_list(x)
}
\arguments{
\item{x}{List.}
}
\description{
Removes \code{NULL} and empty lists from a list recursively.
}
\examples{
x <- list(
  NULL,
  list(),
  list(list(TRUE, list())),
  list(NULL, NULL, list()),
  1,
  list(2, NULL),
  df = data.frame(a = 1, b = 2),
  integer(),
  list(integer())
)
str(trim_list(x))
}
