\name{validPlot}
\alias{validPlot}
\title{Valid Graphs}
\usage{
validPlot(low, up)
}
\arguments{
\item{low}{The lower limit in the graph.}

\item{up}{The upper limit in the graph.}
}
\value{
A graph
}
\description{
This function draws the valid scores
}
\details{
An interactive graphing tool for the cluster validation scores
}
\author{
   Ozan Cinar, Ozlem Ilk-Dag, Cem Iyigun
}