% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{write_table}
\alias{write_table}
\title{Write table to file}
\usage{
write_table(x, file = "")
}
\arguments{
\item{x}{Table.}

\item{file}{Output file name. "" indicates output to the console.}
}
\description{
Writes a table to a standard format.
}
\examples{
df <- data.frame(
  date = Sys.Date(),
  date_time = Sys.time(),
  string = "abc",
  integer = as.integer(123),
  number = 123.456,
  boolean = TRUE
)
write_table(df, "")
}
\seealso{
Other Table functions: \code{\link{as_table}},
  \code{\link{create_table}}, \code{\link{is_table}},
  \code{\link{parse_table_type}},
  \code{\link{parse_table}}, \code{\link{rbind_tables}},
  \code{\link{remove_empty_dimensions}}
}
