% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{geocode}
\alias{geocode}
\title{Geocode Wrapper}
\usage{
geocode(addresses_to_geocode, google_api_key = NA)
}
\arguments{
\item{addresses_to_geocode}{A vector of addresses}

\item{google_api_key}{OPTIONAL API Key}
}
\value{
dataframe
}
\description{
This provides a wrapper for Google's geocoding service
}
