% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{parse_table}
\alias{parse_table}
\title{Parse a table}
\usage{
parse_table(x, functions, as = c("table", "list"), enclos = globalenv(),
  ...)
}
\arguments{
\item{x}{Table.}

\item{functions}{Function or list thereof. If a list, any names are reused in the parsed results.}

\item{as}{Type of object to return. If "table", a table of the same type as \code{x}, otherwise a list.}

\item{enclos}{Environment in which to look for objects not found in \code{x}.}

\item{...}{Arguments passed to \code{\link{feval}}.}
}
\description{
Applies parsers to a table and coerces the result to a table of the same type.
}
\examples{
df <- data.frame(x = 1:5, y = 6:10)
parsers <- list(
  xx = function(x) {x + 10},
  xy = function(x, y) {x + y}
)
parse_table(df, parsers)
parse_table(as.list(df), parsers)
parse_table(df, parsers, as = "list")
parsers$z = function(z) {z * 10}
parse_table(df, parsers)
parse_table(df, parsers, as = "list")
}
\seealso{
Other Table functions: \code{\link{as_table}},
  \code{\link{create_table}}, \code{\link{is_table}},
  \code{\link{parse_table_type}},
  \code{\link{remove_empty_dimensions}}
}
