% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{sample_interval}
\alias{sample_interval}
\title{Sample variables by interval}
\usage{
sample_interval(.data, pos, breaks)
}
\arguments{
\item{.data}{(data.frame) Variables to sample.}

\item{pos}{(numeric)}

\item{breaks}{(numeric) Sorted (weakly) increasingly.}
}
\description{
Sample variables by interval
}
