\name{validManipulate}
\alias{validManipulate}
\title{Interactive Cluster Validation Graphs}
\description{
This function draws the valid scores in an interactive graph. 
}
\usage{
validManipulate(low, up)
}
\arguments{
\item{low}{The lower limit in the graph.}

\item{up}{The upper limit in the graph.}
}
\value{
A graph
}
\details{
An interactive graphing tool for the cluster validation scores
}
\author{
   Ozan Cinar, Ozlem Ilk-Dag, Cem Iyigun
}